+++ person_file.rb
@@ -1,7 +1,7 @@
-# require_relative 'decorator_file'
+require_relative 'decorator_file'
 
 class Person < Nameable
-  # getter & setter
+  # setter
   attr_accessor :name, :age
   # # getter
   attr_reader :id
@@ -28,37 +28,4 @@ class Person < Nameable
     is_of_age? || @parent_permission
   end
   public :can_use_services?
-end
-
-class Nameable
-  def correct_name
-    raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
-  end
-end
-
-class BaseDecorator < Nameable
-  attr_accessor :name
-
-  def initialize(name)
-    super()
-    @name = name
-  end
-
-  def correct_name
-    @name.correct_name
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @name.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    return @name.correct_name[0..9] if @name.correct_name.length > 10
-
-    @name.correct_name
-  end
 end
